# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sports System API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/leagues/{league_id}": {
      "get": {
        "tags": [
          "leagues"
        ],
        "summary": "Get League",
        "operationId": "get_league_leagues__league_id__get",
        "parameters": [
          {
            "name": "league_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "League Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeagueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "leagues"
        ],
        "summary": "Update League",
        "operationId": "update_league_leagues__league_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "league_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "League Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeagueUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeagueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "leagues"
        ],
        "summary": "Archive League",
        "operationId": "archive_league_leagues__league_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "league_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "League Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/leagues/{league_id}/narrative/today": {
      "get": {
        "tags": [
          "narratives"
        ],
        "summary": "Get Today",
        "operationId": "get_today_leagues__league_id__narrative_today_get",
        "parameters": [
          {
            "name": "league_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "League Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/NarrativeResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Today Leagues  League Id  Narrative Today Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/leagues/{league_id}/narrative/{narrative_date}": {
      "get": {
        "tags": [
          "narratives"
        ],
        "summary": "Get For Date",
        "operationId": "get_for_date_leagues__league_id__narrative__narrative_date__get",
        "parameters": [
          {
            "name": "league_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "League Id"
            }
          },
          {
            "name": "narrative_date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Narrative Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NarrativeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/leagues/{league_id}/narrative/generate": {
      "post": {
        "tags": [
          "narratives"
        ],
        "summary": "Generate Narrative",
        "operationId": "generate_narrative_leagues__league_id__narrative_generate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "league_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "League Id"
            }
          },
          {
            "name": "target_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Date"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NarrativeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/leagues/{league_id}/resume/generate": {
      "post": {
        "tags": [
          "narratives"
        ],
        "summary": "Generate Resume",
        "operationId": "generate_resume_leagues__league_id__resume_generate_post",
        "parameters": [
          {
            "name": "league_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "League Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/leagues/{league_id}/ai/generation-history": {
      "get": {
        "tags": [
          "narratives"
        ],
        "summary": "Get Generation History",
        "operationId": "get_generation_history_leagues__league_id__ai_generation_history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "league_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "League Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AIGenerationResponse"
                  },
                  "title": "Response Get Generation History Leagues  League Id  Ai Generation History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIGenerationResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "generation_type": {
            "type": "string",
            "title": "Generation Type"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "generation_type",
          "count",
          "created_at"
        ],
        "title": "AIGenerationResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LeagueResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "created_by_id": {
            "type": "integer",
            "title": "Created By Id"
          },
          "sports_config": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Sports Config"
          },
          "transfer_window_enabled": {
            "type": "boolean",
            "title": "Transfer Window Enabled"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "status": {
            "$ref": "#/components/schemas/LeagueStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "member_count": {
            "type": "integer",
            "title": "Member Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "description",
          "created_by_id",
          "sports_config",
          "transfer_window_enabled",
          "timezone",
          "status",
          "created_at"
        ],
        "title": "LeagueResponse"
      },
      "LeagueStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "ARCHIVED"
        ],
        "title": "LeagueStatus"
      },
      "LeagueUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "sports_config": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sports Config"
          },
          "transfer_window_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transfer Window Enabled"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LeagueStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "LeagueUpdate"
      },
...
      "NarrativeResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "narrative_date": {
            "type": "string",
            "format": "date",
            "title": "Narrative Date"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "narrative_date",
          "content",
          "generated_at"
        ],
        "title": "NarrativeResponse"
      },
...
      "ResumeResponse": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "ResumeResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}